renamed EventFactory > eventMap (in keeping with vuex convention of *Maps)
